Uses Pointer Manipulation to Sort an array of ints and an array of their pointers, and see the difference after each sorting
The program does the following:
Define data types for ... 
-A pointer to an integer
-An array of five integers
-An array of five pointers to integers
Have an initialization function that ... 
--nitializes the elements of an array of five integers to random integers. Use rand to generate the random integers, using srand with the return from getpid to seed the random number generation.
-Initializes the elements of an array of five pointers to integers to point to the corresponding elements of the array of integers. (The structure is anaolgous to the one created in this example that we discussed in class.)
Have a function that prints an array of five integers. 
Have a function that prints the integers pointed to by an array of five pointers to integers.
Have a function that uses a bubble-sort to sort an array of five integers, in ascending order of the integers.
Does not use any random array access
Have a function that uses a bubble-sort to sort an array of five pointers to integers, in ascending order of the integers pointed to by the pointers (i.e., does not sort the integer array - sorts the array of pointers so that when you print it, the integers come out in ascending order). 
Have a main function that ... 
-Declares an array of five integers and an array of five pointers to integers.
-Initializes the arrays (as described above).
-Prints the array of integers.
-Sorts the array of pointers (as described above).
-Prints the integers pointed to by the array of pointers.
-Sorts the array of integers.
-Prints the array of integers.
-Prints the integers pointed to by the array of pointers.